requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
streamlit==1.29.0
pydantic==2.5.2
requests==2.31.0
certifi==2024.2.2 